+++ R/reactives.R
@@ -344,7 +344,7 @@ as.list.reactivevalues <- function(x, all.names=FALSE, ...) {
 
 #' Convert a reactivevalues object to a list
 #'
-#' This function does something similar to what you might \code{\link{as.list}}
+#' This function does something similar to what you might \code{\link[base]{as.list}}
 #' to do. The difference is that the calling context will take dependencies on
 #' every object in the reactivevalues object. To avoid taking dependencies on
 #' all the objects, you can wrap the call with \code{\link{isolate}()}.
@@ -526,6 +526,7 @@ Observable <- R6Class(
       .mostRecentCtxId <<- ctx$id
       ctx$onInvalidate(function() {
         .invalidated <<- TRUE
+        .value <<- NULL # Value can be GC'd, it won't be read once invalidated
         .dependents$invalidate()
       })
       .execCount <<- .execCount + 1L
@@ -1109,7 +1110,7 @@ setAutoflush <- local({
 #' @return A no-parameter function that can be called from a reactive context,
 #'   in order to cause that context to be invalidated the next time the timer
 #'   interval elapses. Calling the returned function also happens to yield the
-#'   current time (as in \code{\link{Sys.time}}).
+#'   current time (as in \code{\link[base]{Sys.time}}).
 #' @seealso \code{\link{invalidateLater}}
 #'
 #' @examples
@@ -1417,7 +1418,7 @@ reactiveFileReader <- function(intervalMillis, session, filePath, readFunc, ...)
 #' The expression given to \code{isolate()} is evaluated in the calling
 #' environment. This means that if you assign a variable inside the
 #' \code{isolate()}, its value will be visible outside of the \code{isolate()}.
-#' If you want to avoid this, you can use \code{\link{local}()} inside the
+#' If you want to avoid this, you can use \code{\link[base]{local}()} inside the
 #' \code{isolate()}.
 #'
 #' This function can also be useful for calling reactive expression at the
@@ -1530,6 +1531,8 @@ maskReactiveContext <- function(expr) {
 #' invalidations that come from its reactive dependencies; it only invalidates
 #' in response to the given event.
 #'
+#' @section \code{ignoreNULL} and \code{ignoreInit}:
+#'
 #' Both \code{observeEvent} and \code{eventReactive} take an \code{ignoreNULL}
 #' parameter that affects behavior when the \code{eventExpr} evaluates to
 #' \code{NULL} (or in the special case of an \code{\link{actionButton}},
@@ -1541,6 +1544,44 @@ maskReactiveContext <- function(expr) {
 #' whereas \code{ignoreNULL=FALSE} is desirable if you want to initially perform
 #' the action/calculation and just let the user re-initiate it (like a
 #' "Recalculate" button).
+#'
+#' Unlike what happens for \code{ignoreNULL}, only \code{observeEvent} takes in an
+#' \code{ignoreInit} argument. By default, \code{observeEvent} will run right when
+#' it is created (except if, at that moment, \code{eventExpr} evaluates to \code{NULL}
+#' and \code{ignoreNULL} is \code{TRUE}). But when responding to a click of an action
+#' button, it may often be useful to set \code{ignoreInit} to \code{TRUE}. For
+#' example, if you're setting up an \code{observeEvent} for a dynamically created
+#' button, then \code{ignoreInit = TRUE} will guarantee that the action (in
+#' \code{handlerExpr}) will only be triggered when the button is actually clicked,
+#' instead of also being triggered when it is created/initialized.
+#'
+#' Even though \code{ignoreNULL} and \code{ignoreInit} can be used for similar
+#' purposes they are independent from one another. Here's the result of combining
+#' these:
+#'
+#' \describe{
+#'   \item{\code{ignoreNULL = TRUE} and \code{ignoreInit = FALSE}}{
+#'      This is the default. This combination means that \code{handlerExpr} will
+#'      run every time that \code{eventExpr} is not \code{NULL}. If, at the time
+#'      of the \code{observeEvent}'s creation, \code{handleExpr} happens to
+#'      \emph{not} be \code{NULL}, then the code runs.
+#'   }
+#'   \item{\code{ignoreNULL = FALSE} and \code{ignoreInit = FALSE}}{
+#'      This combination means that \code{handlerExpr} will run every time no
+#'      matter what.
+#'   }
+#'   \item{\code{ignoreNULL = FALSE} and \code{ignoreInit = TRUE}}{
+#'      This combination means that \code{handlerExpr} will \emph{not} run when
+#'      the \code{observeEvent} is created (because \code{ignoreInit = TRUE}),
+#'      but it will run every other time.
+#'   }
+#'   \item{\code{ignoreNULL = TRUE} and \code{ignoreInit = TRUE}}{
+#'      This combination means that \code{handlerExpr} will \emph{not} run when
+#'      the \code{observeEvent} is created (because \code{ignoreInit = TRUE}).
+#'      After that, \code{handlerExpr} will run every time that \code{eventExpr}
+#'      is not \code{NULL}.
+#'   }
+#' }
 #'
 #' @param eventExpr A (quoted or unquoted) expression that represents the event;
 #'   this can be a simple reactive value like \code{input$click}, a call to a
@@ -1583,6 +1624,15 @@ maskReactiveContext <- function(expr) {
 #' @param ignoreNULL Whether the action should be triggered (or value
 #'   calculated, in the case of \code{eventReactive}) when the input is
 #'   \code{NULL}. See Details.
+#' @param ignoreInit If \code{TRUE}, then, when this \code{observeEvent} is
+#'   first created/initialized, ignore the \code{handlerExpr} (the second
+#'   argument), whether it is otherwise supposed to run or not. The default is
+#'   \code{FALSE}. See Details.
+#' @param once Whether this \code{observeEvent} should be immediately destroyed
+#'   after the first time that the code in \code{handlerExpr} is run. This
+#'   pattern is useful when you want to subscribe to a event that should only
+#'   happen once.
+#'
 #' @return \code{observeEvent} returns an observer reference class object (see
 #'   \code{\link{observe}}). \code{eventReactive} returns a reactive expression
 #'   object (see \code{\link{reactive}}).
@@ -1592,37 +1642,71 @@ maskReactiveContext <- function(expr) {
 #' @examples
 #' ## Only run this example in interactive R sessions
 #' if (interactive()) {
-#'   ui <- fluidPage(
-#'     column(4,
-#'       numericInput("x", "Value", 5),
-#'       br(),
-#'       actionButton("button", "Show")
+#'
+#'   ## App 1: Sample usage
+#'   shinyApp(
+#'     ui = fluidPage(
+#'       column(4,
+#'         numericInput("x", "Value", 5),
+#'         br(),
+#'         actionButton("button", "Show")
+#'       ),
+#'       column(8, tableOutput("table"))
 #'     ),
-#'     column(8, tableOutput("table"))
+#'     server = function(input, output) {
+#'       # Take an action every time button is pressed;
+#'       # here, we just print a message to the console
+#'       observeEvent(input$button, {
+#'         cat("Showing", input$x, "rows\n")
+#'       })
+#'       # Take a reactive dependency on input$button, but
+#'       # not on any of the stuff inside the function
+#'       df <- eventReactive(input$button, {
+#'         head(cars, input$x)
+#'       })
+#'       output$table <- renderTable({
+#'         df()
+#'       })
+#'     }
 #'   )
-#'   server <- function(input, output) {
-#'     # Take an action every time button is pressed;
-#'     # here, we just print a message to the console
-#'     observeEvent(input$button, {
-#'       cat("Showing", input$x, "rows\n")
-#'     })
-#'     # Take a reactive dependency on input$button, but
-#'     # not on any of the stuff inside the function
-#'     df <- eventReactive(input$button, {
-#'       head(cars, input$x)
-#'     })
-#'     output$table <- renderTable({
-#'       df()
-#'     })
-#'   }
-#'   shinyApp(ui=ui, server=server)
+#'
+#'   ## App 2: Using `once`
+#'   shinyApp(
+#'     ui = basicPage( actionButton("go", "Go")),
+#'     server = function(input, output, session) {
+#'       observeEvent(input$go, {
+#'         print(paste("This will only be printed once; all",
+#'               "subsequent button clicks won't do anything"))
+#'       }, once = TRUE)
+#'     }
+#'   )
+#'
+#'   ## App 3: Using `ignoreInit` and `once`
+#'   shinyApp(
+#'     ui = basicPage(actionButton("go", "Go")),
+#'     server = function(input, output, session) {
+#'       observeEvent(input$go, {
+#'         insertUI("#go", "afterEnd",
+#'                  actionButton("dynamic", "click to remove"))
+#'
+#'         # set up an observer that depends on the dynamic
+#'         # input, so that it doesn't run when the input is
+#'         # created, and only runs once after that (since
+#'         # the side effect is remove the input from the DOM)
+#'         observeEvent(input$dynamic, {
+#'           removeUI("#dynamic")
+#'         }, ignoreInit = TRUE, once = TRUE)
+#'       })
+#'     }
+#'   )
 #' }
 #' @export
 observeEvent <- function(eventExpr, handlerExpr,
   event.env = parent.frame(), event.quoted = FALSE,
   handler.env = parent.frame(), handler.quoted = FALSE,
-  label=NULL, suspended=FALSE, priority=0, domain=getDefaultReactiveDomain(),
-  autoDestroy = TRUE, ignoreNULL = TRUE) {
+  label = NULL, suspended = FALSE, priority = 0,
+  domain = getDefaultReactiveDomain(), autoDestroy = TRUE,
+  ignoreNULL = TRUE, ignoreInit = FALSE, once = FALSE) {
 
   eventFunc <- exprToFunction(eventExpr, event.env, event.quoted)
   if (is.null(label))
@@ -1632,16 +1716,29 @@ observeEvent <- function(eventExpr, handlerExpr,
   handlerFunc <- exprToFunction(handlerExpr, handler.env, handler.quoted)
   handlerFunc <- wrapFunctionLabel(handlerFunc, "observeEventHandler", ..stacktraceon = TRUE)
 
-  invisible(observe({
+  initialized <- FALSE
+
+  o <- observe({
     e <- eventFunc()
+
+    if (ignoreInit && !initialized) {
+      initialized <<- TRUE
+      return()
+    }
 
     if (ignoreNULL && isNullEvent(e)) {
       return()
     }
 
+    if (once) {
+      on.exit(o$destroy())
+    }
+
     isolate(handlerFunc())
   }, label = label, suspended = suspended, priority = priority, domain = domain,
-    autoDestroy = TRUE, ..stacktraceon = FALSE))
+  autoDestroy = TRUE, ..stacktraceon = FALSE)
+
+  invisible(o)
 }
 
 #' @rdname observeEvent
@@ -1649,8 +1746,8 @@ observeEvent <- function(eventExpr, handlerExpr,
 eventReactive <- function(eventExpr, valueExpr,
   event.env = parent.frame(), event.quoted = FALSE,
   value.env = parent.frame(), value.quoted = FALSE,
-  label=NULL, domain=getDefaultReactiveDomain(),
-  ignoreNULL = TRUE) {
+  label = NULL, domain = getDefaultReactiveDomain(),
+  ignoreNULL = TRUE, ignoreInit = FALSE) {
 
   eventFunc <- exprToFunction(eventExpr, event.env, event.quoted)
   if (is.null(label))
@@ -1660,13 +1757,17 @@ eventReactive <- function(eventExpr, valueExpr,
   handlerFunc <- exprToFunction(valueExpr, value.env, value.quoted)
   handlerFunc <- wrapFunctionLabel(handlerFunc, "eventReactiveHandler", ..stacktraceon = TRUE)
 
+  initialized <- FALSE
+
   invisible(reactive({
     e <- eventFunc()
 
-    validate(need(
-      !ignoreNULL || !isNullEvent(e),
-      message = FALSE
-    ))
+    if (ignoreInit && !initialized) {
+      initialized <<- TRUE
+      req(FALSE)
+    }
+
+    req(!ignoreNULL || !isNullEvent(e))
 
     isolate(handlerFunc())
   }, label = label, domain = domain, ..stacktraceon = FALSE))
@@ -1674,4 +1775,247 @@ eventReactive <- function(eventExpr, valueExpr,
 
 isNullEvent <- function(value) {
   is.null(value) || (inherits(value, 'shinyActionButtonValue') && value == 0)
+}
+
+#' Slow down a reactive expression with debounce/throttle
+#'
+#' Transforms a reactive expression by preventing its invalidation signals from
+#' being sent unnecessarily often. This lets you ignore a very "chatty" reactive
+#' expression until it becomes idle, which is useful when the intermediate
+#' values don't matter as much as the final value, and the downstream
+#' calculations that depend on the reactive expression take a long time.
+#' \code{debounce} and \code{throttle} use different algorithms for slowing down
+#' invalidation signals; see Details.
+#'
+#' @section Limitations:
+#'
+#'   Because R is single threaded, we can't come close to guaranteeing that the
+#'   timing of debounce/throttle (or any other timing-related functions in
+#'   Shiny) will be consistent or accurate; at the time we want to emit an
+#'   invalidation signal, R may be performing a different task and we have no
+#'   way to interrupt it (nor would we necessarily want to if we could).
+#'   Therefore, it's best to think of the time windows you pass to these
+#'   functions as minimums.
+#'
+#'   You may also see undesirable behavior if the amount of time spent doing
+#'   downstream processing for each change approaches or exceeds the time
+#'   window: in this case, debounce/throttle may not have any effect, as the
+#'   time each subsequent event is considered is already after the time window
+#'   has expired.
+#'
+#' @details
+#'
+#' This is not a true debounce/throttle in that it will not prevent \code{r}
+#' from being called many times (in fact it may be called more times than
+#' usual), but rather, the reactive invalidation signal that is produced by
+#' \code{r} is debounced/throttled instead. Therefore, these functions should be
+#' used when \code{r} is cheap but the things it will trigger (downstream
+#' outputs and reactives) are expensive.
+#'
+#' Debouncing means that every invalidation from \code{r} will be held for the
+#' specified time window. If \code{r} invalidates again within that time window,
+#' then the timer starts over again. This means that as long as invalidations
+#' continually arrive from \code{r} within the time window, the debounced
+#' reactive will not invalidate at all. Only after the invalidations stop (or
+#' slow down sufficiently) will the downstream invalidation be sent.
+#'
+#' \code{ooo-oo-oo---- => -----------o-}
+#'
+#' (In this graphical depiction, each character represents a unit of time, and
+#' the time window is 3 characters.)
+#'
+#' Throttling, on the other hand, delays invalidation if the \emph{throttled}
+#' reactive recently (within the time window) invalidated. New \code{r}
+#' invalidations do not reset the time window. This means that if invalidations
+#' continually come from \code{r} within the time window, the throttled reactive
+#' will invalidate regularly, at a rate equal to or slower than than the time
+#' window.
+#'
+#' \code{ooo-oo-oo---- => o--o--o--o---}
+#'
+#' @param r A reactive expression (that invalidates too often).
+#' @param millis The debounce/throttle time window. You may optionally pass a
+#'   no-arg function or reactive expression instead, e.g. to let the end-user
+#'   control the time window.
+#' @param priority Debounce/throttle is implemented under the hood using
+#'   \link[=observe]{observers}. Use this parameter to set the priority of
+#'   these observers. Generally, this should be higher than the priorities of
+#'   downstream observers and outputs (which default to zero).
+#' @param domain See \link{domains}.
+#'
+#' @examples
+#' ## Only run examples in interactive R sessions
+#' if (interactive()) {
+#' options(device.ask.default = FALSE)
+#'
+#' library(shiny)
+#' library(magrittr)
+#'
+#' ui <- fluidPage(
+#'   plotOutput("plot", click = clickOpts("hover")),
+#'   helpText("Quickly click on the plot above, while watching the result table below:"),
+#'   tableOutput("result")
+#' )
+#'
+#' server <- function(input, output, session) {
+#'   hover <- reactive({
+#'     if (is.null(input$hover))
+#'       list(x = NA, y = NA)
+#'     else
+#'       input$hover
+#'   })
+#'   hover_d <- hover %>% debounce(1000)
+#'   hover_t <- hover %>% throttle(1000)
+#'
+#'   output$plot <- renderPlot({
+#'     plot(cars)
+#'   })
+#'
+#'   output$result <- renderTable({
+#'     data.frame(
+#'       mode = c("raw", "throttle", "debounce"),
+#'       x = c(hover()$x, hover_t()$x, hover_d()$x),
+#'       y = c(hover()$y, hover_t()$y, hover_d()$y)
+#'     )
+#'   })
+#' }
+#'
+#' shinyApp(ui, server)
+#' }
+#'
+#' @export
+debounce <- function(r, millis, priority = 100, domain = getDefaultReactiveDomain()) {
+
+  # TODO: make a nice label for the observer(s)
+
+  force(r)
+  force(millis)
+
+  if (!is.function(millis)) {
+    origMillis <- millis
+    millis <- function() origMillis
+  }
+
+  v <- reactiveValues(
+    trigger = NULL,
+    when = NULL # the deadline for the timer to fire; NULL if not scheduled
+  )
+
+  # Responsible for tracking when f() changes.
+  firstRun <- TRUE
+  observe({
+    r()
+
+    if (firstRun) {
+      # During the first run we don't want to set v$when, as this will kick off
+      # the timer. We only want to do that when we see r() change.
+      firstRun <<- FALSE
+      return()
+    }
+
+    # The value (or possibly millis) changed. Start or reset the timer.
+    v$when <- Sys.time() + millis()/1000
+  }, label = "debounce tracker", domain = domain, priority = priority)
+
+  # This observer is the timer. It rests until v$when elapses, then touches
+  # v$trigger.
+  observe({
+    if (is.null(v$when))
+      return()
+
+    now <- Sys.time()
+    if (now >= v$when) {
+      # Mod by 999999999 to get predictable overflow behavior
+      v$trigger <- isolate(v$trigger %OR% 0) %% 999999999 + 1
+      v$when <- NULL
+    } else {
+      invalidateLater((v$when - now) * 1000)
+    }
+  }, label = "debounce timer", domain = domain, priority = priority)
+
+  # This is the actual reactive that is returned to the user. It returns the
+  # value of r(), but only invalidates/updates when v$trigger is touched.
+  er <- eventReactive(v$trigger, {
+    r()
+  }, label = "debounce result", ignoreNULL = FALSE, domain = domain)
+
+  # Force the value of er to be immediately cached upon creation. It's very hard
+  # to explain why this observer is needed, but if you want to understand, try
+  # commenting it out and studying the unit test failure that results.
+  primer <- observe({
+    primer$destroy()
+    er()
+  }, label = "debounce primer", domain = domain, priority = priority)
+
+  er
+}
+
+#' @rdname debounce
+#' @export
+throttle <- function(r, millis, priority = 100, domain = getDefaultReactiveDomain()) {
+
+  # TODO: make a nice label for the observer(s)
+
+  force(r)
+  force(millis)
+
+  if (!is.function(millis)) {
+    origMillis <- millis
+    millis <- function() origMillis
+  }
+
+  v <- reactiveValues(
+    trigger = 0,
+    lastTriggeredAt = NULL, # Last time we fired; NULL if never
+    pending = FALSE # If TRUE, trigger again when timer elapses
+  )
+
+  blackoutMillisLeft <- function() {
+    if (is.null(v$lastTriggeredAt)) {
+      0
+    } else {
+      max(0, (v$lastTriggeredAt + millis()/1000) - Sys.time()) * 1000
+    }
+  }
+
+  trigger <- function() {
+    v$lastTriggeredAt <- Sys.time()
+    # Mod by 999999999 to get predictable overflow behavior
+    v$trigger <- isolate(v$trigger) %% 999999999 + 1
+    v$pending <- FALSE
+  }
+
+  # Responsible for tracking when f() changes.
+  observeEvent(r(), {
+    if (v$pending) {
+      # In a blackout period and someone already scheduled; do nothing
+    } else if (blackoutMillisLeft() > 0) {
+      # In a blackout period but this is the first change in that period; set
+      # v$pending so that a trigger will be scheduled at the end of the period
+      v$pending <- TRUE
+    } else {
+      # Not in a blackout period. Trigger, which will start a new blackout
+      # period.
+      trigger()
+    }
+  }, label = "throttle tracker", ignoreNULL = FALSE, priority = priority, domain = domain)
+
+  observe({
+    if (!v$pending) {
+      return()
+    }
+
+    timeout <- blackoutMillisLeft()
+    if (timeout > 0) {
+      invalidateLater(timeout)
+    } else {
+      trigger()
+    }
+  }, priority = priority, domain = domain)
+
+  # This is the actual reactive that is returned to the user. It returns the
+  # value of r(), but only invalidates/updates when v$trigger is touched.
+  eventReactive(v$trigger, {
+    r()
+  }, label = "throttle result", ignoreNULL = FALSE, domain = domain)
 }